Make an empty vision periodic (#11)
* Make an empty vision periodic
* ran spotless

The reason we are doing this is that in sim we are seeing lots of overruns in the command
loop due to vision periodic.  We're not sure if it is sim-only or not, but we're not using sim
at this point so reducing the warning/error on the console is a big win.