Copy files swervedrive files from YAGSL-Example, adding the 7660 files.
This is a straight copy of a YAGSL-example branch that I had made
in order to address our problems with our NAVX.

The basic problem is that setting inversion in YAGSL does not change
the gyro to read Counter Clockwise positive as desired.  We have
not yet root-caused this issue, but the crux of the problem is that
the values returned from YAGSL's swervelib.imu.NavXSwerve:getRawRotation3d
do not change when 'inversion' is set.

I believe that the issue is that using 'unaryMinus' is not actually
what you want to do, and will not function correctly if there is
rotation around the X or Y access.  If you want to invert over the
Z access, then converting to a Rotation2d and inverting that is
what you want, or alternatively inverting the getZ() with -1.

These changes were made on top of YAGSL-Example at it's current
'dev' branch a1b981115b84e6a96ca96ed2eca4d58c6784dee1 which is
6 commits ahead of the 2025.7.1 tag.

As to what changes are here and why:

The goal is to replace SwerveSubsystem's swerveDrive (a SwerveDrive
type) with our SwerveDrive7660 which extends
swervelib.swervedrive.SwerveDrive In order to accomplish the goal, we
will need to create SwerveParser7660 which extends
swervelib.SwerveParser.

Our SwerveSubsystem has a change to line 90 which calls the
SwerveParser7660.

   swerveDrive = new SwerveParser7660(directory).createSwerveDrive(Constants.MAX_SPEED, startingPose);

The only point of SwerveParser7660 is to have it create for us a
SwerveDrive7660. It has only 2 functions:
 1. A constructor SwerveParser(File directory)
    this only needs to call super(directory)
    it has to be here because you have to call a superclass's constructor

 2. createSwerveDrive
    this has to be a copy of SwerveParser's version with the only
    change being that it will return new SwerveDrive7660

SwerveDrive7660 extends swervelib.swervedrive.SwerveDrive
It overrides one function in it's superclass.

   public Rotation2d getYaw() {
    return Rotation2d.fromRadians(-imuReadingCache.getValue().getZ());
   }

So this is how we are forcing the Z (Yaw) axis inversion to get
Counter Clockwise to read positive.